% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swe_is_repealed.R
\name{swe_is_repealed}
\alias{swe_is_repealed}
\title{Check whether a Swedish statute has been repealed}
\usage{
swe_is_repealed(sfs_nr)
}
\arguments{
\item{sfs_nr}{Character. SFS number in the format "1974:152".}
}
\value{
Logical. \code{TRUE} if repealed, \code{FALSE} otherwise.
}
\description{
A thin wrapper around \code{\link[=swe_get_doc]{swe_get_doc()}}.
}
